// src/services/model/getProductLineSlug200Data.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * TrinityKitWP API
 * API para criar posts e fazer upload de imagens no WordPress
 * OpenAPI spec version: 1.0.0
 */
import type { GetProductLineSlug200DataParent } from './getProductLineSlug200DataParent';
import type { GetProductLineSlug200DataSubcategoriesItem } from './getProductLineSlug200DataSubcategoriesItem';

export type GetProductLineSlug200Data = {
  parent?: GetProductLineSlug200DataParent;
  subcategories?: GetProductLineSlug200DataSubcategoriesItem[];
};