// src/services/model/getProductLineSlug200DataSubcategoriesItem.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * TrinityKitWP API
 * API para criar posts e fazer upload de imagens no WordPress
 * OpenAPI spec version: 1.0.0
 */
import type { GetProductLineSlug200DataSubcategoriesItemProductsItem } from './getProductLineSlug200DataSubcategoriesItemProductsItem';

export type GetProductLineSlug200DataSubcategoriesItem = {
  /** Nome da subcategoria */
  name?: string;
  /** Slug da subcategoria */
  slug?: string;
  products?: GetProductLineSlug200DataSubcategoriesItemProductsItem[];
};